+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## **unreleased**
+
+- Style
+  - Omit redundant table headers
+
 ## 0.0.1
 
 - Added